pantera_bot/requirements.txt
python-telegram-bot==20.7
openai==1.14.3
python-dotenv==1.0.1
requests==2.31.0
pytz
